chore: file cleanups
