+++ inc/Modules/Core/Assets.php
@@ -129,6 +129,8 @@ class Assets implements Registrable {
 	public function defer_scripts( string $tag, string $handle ): string {
 		$defer_handles = [
 			self::SETTINGS_SCRIPT_HANDLE,
+			self::PLUGIN_MANAGER_SCRIPT_HANDLE,
+			self::PULL_REQUESTS_SCRIPT_HANDLE,
 		];
 
 		// Bail if we don't need to defer.
